Patially migrated to python3
* changed structure for libraries and code
* rewrote queries without ORM
* reworked config file
* added debug options for current development steps
