+++ app/__init__.py
@@ -1,16 +1,18 @@
-from app import config
 from flask import Flask, Blueprint
-from sqlalchemy import create_engine
+from app.reports.views import reports
+from app.views import statics
 import chartkick
+import config
 
-app = Flask(__name__)
-app.config.from_object('config')
-
-db = create_engine(config.DB_URI, echo=True)
-
-custom_path = config.CUSTOM_PATH
+webapp = Flask(__name__)
+webapp.config.from_object('config')
 
 ck = Blueprint('ck_page', __name__, static_folder=chartkick.js(), static_url_path='/static')
-app.register_blueprint(ck, url_prefix='/ck')
-app.jinja_env.add_extension("chartkick.ext.charts")
-app.jinja_env.add_extension('jinja2.ext.do')
+
+webapp.register_blueprint(ck, url_prefix='/ck')
+webapp.jinja_env.add_extension("chartkick.ext.charts")
+webapp.jinja_env.add_extension('jinja2.ext.do')
+
+# Adding routes
+webapp.register_blueprint(statics)
+webapp.register_blueprint(reports)
