--- app/config.py.example
@@ -1,3 +1,0 @@
-DB_URI = 'postgresql://<URI_TO_YOUR_DATABASE>?client_encoding=utf8'
-SQLALCHEMY_TRACK_MODIFICATIONS = True
-CUSTOM_PATH='<some_path>'
