+++ app/db.py
@@ -1,0 +1,4 @@
+from sqlalchemy import create_engine
+import config
+
+db = create_engine(config.DB_URI, echo=True)
