+++ app/static/jquery.address.js
@@ -37,7 +37,7 @@
                 return (_h.pushState && _opts.state !== UNDEFINED);
             },
             _hrefState = function() {
-                return ('/' + _l.pathname.replace(new RegExp(_opts.state), '') + 
+                return ('/' + _l.pathname.replace(new RegExp(_opts.state), '') +
                     _l.search + (_hrefHash() ? '#' + _hrefHash() : '')).replace(_re, '/');
             },
             _hrefHash = function() {
@@ -49,9 +49,9 @@
             },
             _window = function() {
                 try {
-                    return top.document !== UNDEFINED && top.document.title !== UNDEFINED && top.jQuery !== UNDEFINED && 
+                    return top.document !== UNDEFINED && top.document.title !== UNDEFINED && top.jQuery !== UNDEFINED &&
                         top.jQuery.address !== UNDEFINED && top.jQuery.address.frames() !== false ? top : window;
-                } catch (e) { 
+                } catch (e) {
                     return window;
                 }
             },
@@ -84,13 +84,13 @@
             },
             _update = function(internal) {
                 _st(_track, 10);
-                return _trigger(CHANGE).isDefaultPrevented() || 
+                return _trigger(CHANGE).isDefaultPrevented() ||
                     _trigger(internal ? INTERNAL_CHANGE : EXTERNAL_CHANGE).isDefaultPrevented();
             },
             _track = function() {
                 if (_opts.tracker !== 'null' && _opts.tracker !== NULL) {
                     var fn = $.isFunction(_opts.tracker) ? _opts.tracker : _t[_opts.tracker],
-                        value = (_l.pathname + _l.search + 
+                        value = (_l.pathname + _l.search +
                                 ($.address && !_supportsState() ? $.address.value() : ''))
                                 .replace(/\/\//, '/').replace(/^\/$/, '');
                     if ($.isFunction(fn)) {
@@ -105,9 +105,9 @@
                 }
             },
             _html = function() {
-                var src = _js() + ':' + FALSE + ';document.open();document.writeln(\'<html><head><title>' + 
-                    _d.title.replace(/\'/g, '\\\'') + '</title><script>var ' + ID + ' = "' + encodeURIComponent(_href()).replace(/\'/g, '\\\'') + 
-                    (_d.domain != _l.hostname ? '";document.domain="' + _d.domain : '') + 
+                var src = _js() + ':' + FALSE + ';document.open();document.writeln(\'<html><head><title>' +
+                    _d.title.replace(/\'/g, '\\\'') + '</title><script>var ' + ID + ' = "' + encodeURIComponent(_href()).replace(/\'/g, '\\\'') +
+                    (_d.domain != _l.hostname ? '";document.domain="' + _d.domain : '') +
                     '";</' + 'script></head></html>\');document.close();';
                 if (_version < 7) {
                     _frame.src = src;
@@ -139,13 +139,13 @@
                     if (_opts.wrap) {
                         var body = $('body'),
                             wrap = $('body > *')
-                                .wrapAll('<div style="padding:' + 
-                                    (_cssint(body, 'marginTop') + _cssint(body, 'paddingTop')) + 'px ' + 
-                                    (_cssint(body, 'marginRight') + _cssint(body, 'paddingRight')) + 'px ' + 
-                                    (_cssint(body, 'marginBottom') + _cssint(body, 'paddingBottom')) + 'px ' + 
+                                .wrapAll('<div style="padding:' +
+                                    (_cssint(body, 'marginTop') + _cssint(body, 'paddingTop')) + 'px ' +
+                                    (_cssint(body, 'marginRight') + _cssint(body, 'paddingRight')) + 'px ' +
+                                    (_cssint(body, 'marginBottom') + _cssint(body, 'paddingBottom')) + 'px ' +
                                     (_cssint(body, 'marginLeft') + _cssint(body, 'paddingLeft')) + 'px;" />')
                                 .parent()
-                                .wrap('<div id="' + ID + '" style="height:100%;overflow:auto;position:relative;' + 
+                                .wrap('<div id="' + ID + '" style="height:100%;overflow:auto;position:relative;' +
                                     (_webkit && !window.statusbar.visible ? 'resize:both;' : '') + '" />');
                         $('html, body')
                             .css({
@@ -261,8 +261,8 @@
             TRUE = true,
             FALSE = false,
             _opts = {
-                autoUpdate: TRUE, 
-                history: TRUE, 
+                autoUpdate: TRUE,
+                history: TRUE,
                 strict: TRUE,
                 frames: TRUE,
                 wrap: FALSE
@@ -273,7 +273,7 @@
             _msie = !$.support.opacity,
             _t = _window(),
             _d = _t.document,
-            _h = _t.history, 
+            _h = _t.history,
             _l = _t.location,
             _si = setInterval,
             _st = setTimeout,
@@ -284,14 +284,14 @@
             _form,
             _url = $('script:last').attr('src'),
             _qi = _url ? _url.indexOf('?') : -1,
-            _title = _d.title, 
+            _title = _d.title,
             _silent = FALSE,
             _loaded = FALSE,
             _juststart = TRUE,
             _updating = FALSE,
-            _listeners = {}, 
+            _listeners = {},
             _value = _href();
-            
+
         if (_msie) {
             _version = parseFloat(_agent.substr(_agent.indexOf('MSIE') + 4));
             if (_d.documentMode && _d.documentMode != _version) {
@@ -307,7 +307,7 @@
                 }
             };
         }
-        
+
         if (_h.navigationMode) {
             _h.navigationMode = 'compatible';
         }
@@ -393,7 +393,7 @@
                     return this;
                 }
                 return _opts.frames;
-            },            
+            },
             strict: function(value) {
                 if (value !== UNDEFINED) {
                     _opts.strict = value;
@@ -449,7 +449,7 @@
                             return this;
                         }
                         if (_supportsState()) {
-                            _h[_opts.history ? 'pushState' : 'replaceState']({}, '', 
+                            _h[_opts.history ? 'pushState' : 'replaceState']({}, '',
                                     _opts.state.replace(/\/$/, '') + (_value === '' ? '/' : _value));
                         } else {
                             _silent = TRUE;
@@ -522,7 +522,7 @@
                             v = [v];
                         }
                         if (n == name) {
-                            v = (value === NULL || value === '') ? [] : 
+                            v = (value === NULL || value === '') ? [] :
                                 (append ? v.concat([value]) : [value]);
                         }
                         for (var j = 0; j < v.length; j++) {
@@ -570,11 +570,11 @@
                     return this;
                 }
                 var arr = _value.split('#');
-                return arr.slice(1, arr.length).join('#');                
+                return arr.slice(1, arr.length).join('#');
             }
         };
     })();
-    
+
     $.fn.address = function(fn) {
         $(this).each(function(index) {
             if (!$(this).data('address')) {
@@ -585,10 +585,10 @@
                     var target = e.currentTarget;
                     if ($(target).is('a')) {
                         e.preventDefault();
-                        var value = fn ? fn.call(target) : 
-                            /address:/.test($(target).attr('rel')) ? $(target).attr('rel').split('address:')[1].split(' ')[0] : 
-                            $.address.state() !== undefined && !/^\/?$/.test($.address.state()) ? 
-                                    $(target).attr('href').replace(new RegExp('^(.*' + $.address.state() + '|\\.)'), '') : 
+                        var value = fn ? fn.call(target) :
+                            /address:/.test($(target).attr('rel')) ? $(target).attr('rel').split('address:')[1].split(' ')[0] :
+                            $.address.state() !== undefined && !/^\/?$/.test($.address.state()) ?
+                                    $(target).attr('href').replace(new RegExp('^(.*' + $.address.state() + '|\\.)'), '') :
                                     $(target).attr('href').replace(/^(#\!?|\.)/, '');
                         $.address.value(value);
                     }
@@ -597,7 +597,7 @@
                     if ($(target).is('form')) {
                         e.preventDefault();
                         var action = $(target).attr('action'),
-                            value = fn ? fn.call(target) : (action.indexOf('?') != -1 ? action.replace(/&$/, '') : action + '?') + 
+                            value = fn ? fn.call(target) : (action.indexOf('?') != -1 ? action.replace(/&$/, '') : action + '?') +
                                 $(target).serialize();
                         $.address.value(value);
                     }
@@ -606,5 +606,5 @@
         });
         return this;
     };
-    
+
 })(jQuery);
