+++ app/templates/about.html
@@ -15,5 +15,36 @@
     </div>
   </div>
 </div>
+<div class="ui vertical stripe segment">
+  <div class="ui middle aligned stackable grid container">
+    <div class="eight wide column">
+      <h4>Links:</h4>
+      <div class="ui divided selection list">
+        <a class="item" href="https://github.com/l13t/pyWBacula/wiki">
+          <div class="ui teal horizontal label">Wiki</div>
+          Documentation page
+        </a>
+        <a class="item" href="https://github.com/l13t/pyWBacula">
+          <div class="ui teal horizontal label"><i class="github icon"></i>GitHub</div>
+          Source codes and issue tracker
+        </a>
+      </div>
+    </div>
+    <div class="eight wide column">
+      <h4>Contacts:</h4>
+      <div class="ui divided selection list">
+        <a class="item" href="mailto:liet@liet.kiev.ua">
+          <div class="ui teal horizontal label"><i class="mail outline icon"></i>Write email</div>
+          Dmytro Prokhorenkov (liet@liet.kiev.ua)
+        </a>
+        <a class="item" href="https://github.com/l13t/pyWBacula/issues">
+          <div class="ui teal horizontal label"><i class="bug icon"></i>BugTracker</div>
+          Write about problems or suggestions
+        </a>
+      </div>
+    </div>
+  </div>
+  </div>
+</div>
 
 {%- include 'footer.html' %}
