+++ app/templates/footer.html
@@ -7,10 +7,7 @@
         </h4>
         <div class="ui inverted link list">
           <a href="/about" class="item">
-            About project
-          </a>
-          <a href="/contact_us" class="item">
-            Contact us
+            About
           </a>
         </div>
       </div>
@@ -22,7 +19,7 @@
         </h4>
         <div class="ui inverted link list">
           <p class="item">
-            version 0.1
+            version 0.2
           </p>
         </div>
       </div>
