+++ app/templates/old_volumes.html
@@ -1,7 +1,7 @@
 {%- include 'header.html' %}
 {%- include 'tables.html' %}
 
-<script src="/static/dist/components/tab.js"></script>
+<script src="https://cdn.jsdelivr.net/npm/semantic-ui@2.4.2/dist/components/tab.js"></script>
 <script src="/static/sinon.js"></script>
 <script type="text/javascript">// <![CDATA[
 $( document ).ready(function() {
