+++ libs/bacula_base64.py
