+++ libs/static_vars.py
