+++ run.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
 __credits__ = ["Dmytro Prokhorenkov"]
@@ -7,10 +7,13 @@ __maintainer__ = "Dmytro Prokhorenkov"
 __email__ = "liet@liet.kiev.ua"
 __status__ = "Development"
 
-from app import app
+from app import webapp
+from flask_debug import Debug
 import os
 
 if __name__ == '__main__':
-    DEBUG = os.getenv("PWB_DEBUG", default="False")
+    DEBUG = bool(os.getenv("PWB_DEBUG", default="False"))
     PORT = os.getenv("PWB_PORT", default="15995")
-    app.run(host='0.0.0.0', port=PORT, debug=DEBUG)
+    Debug(webapp)
+    webapp.config['FLASK_DEBUG_DISABLE_STRICT'] = True
+    webapp.run(host='0.0.0.0', port=PORT, debug=DEBUG, use_debugger=DEBUG, use_reloader=DEBUG)
